Cast parameter passed to ctype(3) functions to uint8 to enforce the compiler to look at the correct piece of memory for the bit mask. Fixes some international alphabet problems.

+++ ggi/mode.c
@@ -1,4 +1,4 @@
-/* $Id: mode.c,v 1.9 2004/10/14 11:01:33 cegger Exp $
+/* $Id: mode.c,v 1.10 2004/10/29 22:52:52 cegger Exp $
 ******************************************************************************
 
    LibGGI Mode management.
@@ -441,15 +441,15 @@ int ggiFPrintMode(FILE *s, ggi_mode *m)
  * for failure.
  */
 
-#define SKIPSPACE   while ((*s) && (isspace((int)*s) || (*s == '.'))) s++
+#define SKIPSPACE   while ((*s) && (isspace((uint8)*s) || (*s == '.'))) s++
  
 /* scan the integer from the string pointer s */
 
 #define SCANINT(x,def)                             \
 	SKIPSPACE;  x = def;                       \
-	if (isdigit((int) *s)) {                   \
+	if (isdigit((uint8) *s)) {                 \
 		x = *s++ - '0';                    \
-		while (isdigit((int) *s)) {        \
+		while (isdigit((uint8) *s)) {      \
 			x = x*10 + (*s++ - '0');   \
 		}                                  \
 	}                                          \
@@ -473,62 +473,62 @@ int ggiParseMode(const char * s, ggi_mode * m)
 
 		SKIPSPACE;
 
-		if ((tolower((int)*s)=='s') || isdigit((int)*s)) {	/* visible */
-			if (! isdigit((int)*s)) s++;
+		if ((tolower((uint8)*s)=='s') || isdigit((uint8)*s)) {	/* visible */
+			if (! isdigit((uint8)*s)) s++;
 			SCANINT(m->visible.x, GGI_AUTO);
-			if (tolower((int) *s) == 'x') {
+			if (tolower((uint8) *s) == 'x') {
 				s++; SCANINT(m->visible.y, GGI_AUTO);
 			}
-			if (tolower((int) *s) == 'x') {
+			if (tolower((uint8) *s) == 'x') {
 				s++; SCANINT(depth, GT_AUTO);
 				GT_SETDEPTH(m->graphtype, depth);
 			}
 			continue;
 		}
 		
-		if ((*s=='#') || (tolower((int) *s)=='v')) {	/* virtual */
+		if ((*s=='#') || (tolower((uint8) *s)=='v')) {	/* virtual */
 			s++; SCANINT(m->virt.x, GGI_AUTO);
-			if (tolower((int) *s) == 'x') {
+			if (tolower((uint8) *s) == 'x') {
 				s++; SCANINT(m->virt.y, GGI_AUTO);
 			}
 			continue;
 		}
 
-		if (tolower((int) *s)=='f') {	/* frames */
+		if (tolower((uint8) *s)=='f') {	/* frames */
 			s++; SCANINT(m->frames, GGI_AUTO);
 			continue;
 		}
 
-		if (tolower((int) *s)=='d') {  /* dpp */
+		if (tolower((uint8) *s)=='d') {  /* dpp */
 			s++; SCANINT(m->dpp.x, GGI_AUTO);
-			if (tolower((int) *s) == 'x') {
+			if (tolower((uint8) *s) == 'x') {
 				s++; SCANINT(m->dpp.y, GGI_AUTO);
 			}
 			continue;
 		}
 
-		if (tolower((int) *s)=='p') {  /* palette */
+		if (tolower((uint8) *s)=='p') {  /* palette */
 			s++; SCANINT(depth, GT_AUTO);
 			GT_SETSCHEME(m->graphtype, GT_PALETTE);
 			GT_SETDEPTH(m->graphtype, depth);
 			continue;
 		}
 
-		if (tolower((int) *s)=='c') {  /* truecolor */
+		if (tolower((uint8) *s)=='c') {  /* truecolor */
 			s++; SCANINT(depth, GT_AUTO);
 			GT_SETSCHEME(m->graphtype, GT_TRUECOLOR);
 			GT_SETDEPTH(m->graphtype, depth);
 			continue;
 		}
 
-		if (tolower((int) *s)=='k') {  /* greyscale */
+		if (tolower((uint8) *s)=='k') {  /* greyscale */
 			s++; SCANINT(depth, GT_AUTO);
 			GT_SETSCHEME(m->graphtype, GT_GREYSCALE);
 			GT_SETDEPTH(m->graphtype, depth);
 			continue;
 		}
 
-		if (tolower((int) *s)=='t') {  /* text */
+		if (tolower((uint8) *s)=='t') {  /* text */
 			s++; SCANINT(depth, GT_AUTO);
 			GT_SETSCHEME(m->graphtype, GT_TEXT);
 			GT_SETDEPTH(m->graphtype, depth);
@@ -552,19 +552,19 @@ int ggiParseMode(const char * s, ggi_mode * m)
 		CHECKGTMODE("GT_TEXT16]", 10, GT_TEXT16);
 		CHECKGTMODE("GT_TEXT32]", 10, GT_TEXT32);
 	
-		if (tolower((int) *s) == 't') {  /* text */
+		if (tolower((uint8) *s) == 't') {  /* text */
 			GT_SETSCHEME(m->graphtype, GT_TEXT);
 			s++;
 		} else
-		if (tolower((int) *s) == 'p') {  /* palette */
+		if (tolower((uint8) *s) == 'p') {  /* palette */
 			GT_SETSCHEME(m->graphtype, GT_PALETTE);
 			s++;
 		} else
-		if (tolower((int) *s) == 'c') {  /* truecolor */
+		if (tolower((uint8) *s) == 'c') {  /* truecolor */
 			GT_SETSCHEME(m->graphtype, GT_TRUECOLOR);
 			s++;
 		} else
-		if (tolower((int) *s) == 'k') {  /* greyscale */
+		if (tolower((uint8) *s) == 'k') {  /* greyscale */
 			GT_SETSCHEME(m->graphtype, GT_GREYSCALE);
 			s++;
 		}
@@ -611,13 +611,13 @@ int ggiParseMode(const char * s, ggi_mode * m)
 	int yposition=0;
 	GGIDPRINT_CORE("ggiParseMode(%p, %p) called\n", s, m);
 	*m = _ggiDefaultMode;
-#define SKIPSPACE   while ( (*s!='\000') && (isspace((int)*s)) ) s++;
+#define SKIPSPACE   while ( (*s!='\000') && (isspace((uint8)*s)) ) s++;
 /* scan the integer from the string pointer s */
 #define SCANINT(x)  SKIPSPACE;                 \
-		    if (isdigit((int)*s)){          \
+		    if (isdigit((uint8)*s)){        \
 		       x=*s-'0';               \
 	               s++;                    \
-		       while (isdigit((int)*s)){    \
+		       while (isdigit((uint8)*s)){  \
 			  x = x*10+ ((int)*s) -'0'; \
 			  s++;                 \
 		       }                       \
@@ -625,27 +625,27 @@ int ggiParseMode(const char * s, ggi_mode * m)
 		    SKIPSPACE
 	/* first number is visible-x: */
 	SCANINT(m->visible.x);
-	if (tolower((int)*s) == 'x') { /* now for the y */
+	if (tolower((uint8)*s) == 'x') { /* now for the y */
 		s++;
 		SCANINT(m->visible.y);
 	}
 	if (*s == '#'){ /* virtual starts here */
 		s++;
 		SCANINT(m->virt.x);
-		if (tolower((int)*s) == 'x') { /* now for the y */
+		if (tolower((uint8)*s) == 'x') { /* now for the y */
 			s++;
 			SCANINT(m->virt.y);
 		}
 	}
-	if (tolower((int)*s) == 'd') { /* dpp starts here */
+	if (tolower((uint8)*s) == 'd') { /* dpp starts here */
 		s++;
 		SCANINT(m->dpp.x);
-		if (tolower((int)*s) == 'x') { /* now for the y */
+		if (tolower((uint8)*s) == 'x') { /* now for the y */
 			s++;
 			SCANINT(m->dpp.y);
 		}
 	}
-	if (tolower((int)*s) == 'f') { /* frames starts here */
+	if (tolower((uint8)*s) == 'f') { /* frames starts here */
 		s++;
 		SCANINT(m->frames);
 	}
@@ -666,19 +666,19 @@ int ggiParseMode(const char * s, ggi_mode * m)
 		CHECKGTMODE("GT_TEXT32]", 10, GT_TEXT32) else
 		{
 			/* scheme */
-			if (tolower((int)*s) == 't') {  /* text mode */
+			if (tolower((uint8)*s) == 't') {  /* text mode */
 				GT_SETSCHEME(m->graphtype, GT_TEXT);
 				s++;
 			} else
-			if (tolower((int)*s) == 'p') {  /* palette mode */
+			if (tolower((uint8)*s) == 'p') {  /* palette mode */
 				GT_SETSCHEME(m->graphtype, GT_PALETTE);
 				s++;
 			} else
-			if (tolower((int)*s) == 'c') {  /* truecolor mode */
+			if (tolower((uint8)*s) == 'c') {  /* truecolor mode */
 				GT_SETSCHEME(m->graphtype, GT_TRUECOLOR);
 				s++;
 			} else
-			if (tolower((int)*s) == 'k') {  /* greyscale mode */
+			if (tolower((uint8)*s) == 'k') {  /* greyscale mode */
 				GT_SETSCHEME(m->graphtype, GT_GREYSCALE);
 				s++;
 			}
